Put application response models into a separate file.
